backend/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
Cython==0.29.28
Django==3.2.14
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
elasticsearch==7.17.4
filelock==3.7.1
gensim==4.2.0
hazm==0.7.0
huggingface-hub==0.8.1
idna==3.3
importlib-metadata==4.12.0
joblib==1.1.0
Markdown==3.4.1
nltk==3.3
numpy==1.21.6
packaging==21.3
pandas==1.3.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.7.9
requests==2.28.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
smart-open==6.0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.12.0
tqdm==4.64.0
transformers==4.20.1
typing_extensions==4.3.0
urllib3==1.26.10
zipp==3.8.1
